requirements.txt
pandas==0.23.4
pyserial==3.4
matplotlib==3.0.2
openpyxl==2.6.2
